+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+[1.4.0] - 2019-10-22
+* Fix minDate, maxDate, disableDates and value (gotoDate) of input now respects format parameter.
+
+[1.3.6] - 2019-10-22
+* New options `onMonthsChange` and `onYearsChange` - Events fire when months or years select changed.
+    `onMonthsChange` throws zero based number of month. Ex: January = 0
+    `onYearsChange` throws year. Ex: 2019
+    Implementation of #59
+    
+[1.3.5] - 2019-10-22
+* New option `weekdayStyle` - Weekday display style added.
+    Possible values are 'long', 'short' or 'narrow'.
+    Implementation of #91
+
 [1.3.4] - 2019-03-30
 * fixed bug #51